requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1